refactor(API/addresses)

+++ api/src/controllers/addresses.controller.js
@@ -1,36 +1,85 @@
 const addressService = require('../services/userAddresses.service')
 const usersService = require('../services/users.service.js')
 
-async function create (req, res, next) {
+async function create (req, res) {
   const user = await usersService.getUserByEmail(req.user.email)
   const newAddress = req.body
   newAddress.userId = user.id
   const requiredData = ['postalCode', 'state', 'city', 'streetName', 'houseNumber']
   let validationErrors = 'The following mandatory data is missing in your request: '
   const errorLength = validationErrors.length
-
   for (const column of requiredData) {
     newAddress[column] || (validationErrors += column + ' ')
   }
-
   if (validationErrors.length > errorLength) return res.status(400).json({ error: validationErrors })
 
   try {
     const wasCreated = await addressService.createAddress(newAddress)
-    wasCreated ? res.send({ message: 'Address created correctly' }) : res.status(400).json({ error: 'Address was not created, it already exists' })
+    const allUserAddresses = await addressService.getUserAddresses(user.id)
+    wasCreated
+      ? res.send({
+        payload: allUserAddresses,
+        message: 'Se guardo la nueva dirección con éxito'
+      })
+      : res.status(400).json({
+        payload: allUserAddresses,
+        error: 'Ya existe la direccion ingresada'
+      })
+  } catch (error) {
+    res.status(400).send(error)
+  }
+}
+
+async function remove (req, res) {
+  const { addressId } = req.params
+  if (!addressId) {
+    return res.status(400).json({ error: 'Must provide the address ID to be deleted' })
+  }
+  if (isNaN(parseInt(addressId))) {
+    return res.status(400).json({ error: 'Invalid address ID' })
+  }
+
+  try {
+    const user = await usersService.getUserByEmail(req.user.email)
+    const wasDeleted = await addressService.removeAddress(addressId)
+    const allUserAddresses = await addressService.getUserAddresses(user.id)
+    wasDeleted
+      ? res.send({
+        payload: allUserAddresses,
+        message: 'Se elimino la dirección con éxito'
+      })
+      : res.status(400).json({
+        payload: allUserAddresses,
+        error: 'No se pudo eliminar la direccion'
+      })
   } catch (error) {
     res.send(error)
   }
 }
 
-async function remove (req, res, next) {
+async function update (req, res) {
   const { addressId } = req.params
   if (!addressId) {
-    return res.status(400).json({ error: 'Must provide the address ID to be deleted' })
+    return res.status(400).json({ error: 'Must provide the address ID to be updated' })
   }
+  if (isNaN(parseInt(addressId))) {
+    return res.status(400).json({ error: 'Invalid address ID' })
+  }
+  const updatedAddress = req.body
+  updatedAddress.id = addressId
   try {
-    const wasDeleted = await addressService.removeAddress(addressId)
-    wasDeleted ? res.json({ message: 'Address deleted correctly' }) : res.status(400).json({ error: 'Address could not be deleted' })
+    const user = await usersService.getUserByEmail(req.user.email)
+    const wasUpdated = await addressService.updateAddress(updatedAddress)
+    const allUserAddresses = await addressService.getUserAddresses(user.id)
+    wasUpdated
+      ? res.send({
+        payload: allUserAddresses,
+        message: 'Se actualizó la dirección con éxito'
+      })
+      : res.status(400).json({
+        payload: allUserAddresses,
+        error: 'No se pudo actualizar la dirección'
+      })
   } catch (error) {
     res.send(error)
   }
@@ -38,5 +87,7 @@ async function remove (req, res, next) {
 
 module.exports = {
   create,
-  remove
+  get,
+  remove,
+  update
 }
